git-svn-id: https://svn.lcsr.jhu.edu/cisst/branches/2011-12-15-fault@3967 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ cisst/cisstVector/tests/vctDataFunctionsDynamicMatrixTest.cpp
@@ -19,6 +19,7 @@ http://www.cisst.org/cisst/license.txt.
 --- end cisst license ---
 */
 
+#include "vctGenericContainerTest.h"
 #include "vctDataFunctionsDynamicMatrixTest.h"
 
 #include <cisstVector/vctDynamicMatrix.h>
@@ -150,7 +151,7 @@ void vctDataFunctionsDynamicMatrixTest::TestScalar(void)
     position = 0;
     for (row = 0; row < mInt.rows(); ++row) {
         for (col = 0; col < mInt.cols(); ++col) {
-            mInt.Element(row, col) = row * 10 + col;
+            mInt.Element(row, col) = static_cast<int>(row * 10 + col);
             CPPUNIT_ASSERT_EQUAL(static_cast<double>(row * 10 + col), cmnDataScalar(mInt, position));
             position++;
         }
@@ -161,7 +162,7 @@ void vctDataFunctionsDynamicMatrixTest::TestScalar(void)
         for (col = 0; col < mmDouble.cols(); ++col) {
             for (subRow = 0; subRow < mmDouble.Element(row, col).rows(); ++subRow) {
                 for (subCol = 0; subCol < mmDouble.Element(row, col).cols(); ++subCol) {
-                    mmDouble.Element(row, col).Element(subRow, subCol) = row * col * 100 + subRow * subCol;
+                    mmDouble.Element(row, col).Element(subRow, subCol) = static_cast<double>(row * col * 100 + subRow * subCol);
                     CPPUNIT_ASSERT_EQUAL(static_cast<double>(row * col * 100 + subRow * subCol),
                                          cmnDataScalar(mmDouble, position));
                     position++;
